Working Core 0.12.0
- Fixed saving bugs
- Fixed menu bugs
- Fixed loading bugs
- Patched up player attack method
- Started fixing enemy bugs

+++ Assets/Scripts/Game/DataSave.cs
@@ -21,7 +21,7 @@ public class DataSave
 
     public Dictionary<string, SaveLevel> Levels = new();
     public string lastLevelID;
-    public int lastX, lastY;
+    public float lastX, lastY;
 
     public Dictionary<string, SaveNPCData> NPCs = new();
 }
